removed the deprecated slack notifier
